web_deprecated/requirements.txt
alembic==1.0.0
appdirs==1.4.4
atomicwrites==1.1.5
attrs==18.1.0
autopep8==1.5.5
bcrypt==3.1.4
certifi==2018.4.16
cffi==1.14.5
chardet==3.0.4
click==6.7
coverage==4.0.3
distlib==0.3.1
filelock==3.0.12
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.1
marshmallow==2.15.3
more-itertools==4.2.0
pipenv==2020.11.15
pluggy==0.7.1
psycopg2==2.8.6
py==1.5.4
pycodestyle==2.6.0
pycparser==2.18
PyJWT==1.6.4
pytest==3.6.4
pytest-cov==2.5.1
python-coveralls==2.9.1
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
PyYAML==3.13
requests==2.19.1
six==1.11.0
SQLAlchemy==1.3.23
toml==0.10.2
urllib3==1.23
virtualenv==20.4.2
virtualenv-clone==0.5.4
Werkzeug==0.14.1
